Uploaded main files of project
Uploading main files for this project:
lru.c implements a doubly linked list that is used as a queue for the
replacement policy.
cache.c implements the cache and all its functions.
test.c contains tests for all the functions in cache.c and lru.c.
Makefile was used to link all the files together.